backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
openai
passlib[bcrypt]
python-dotenv
pyjwt
pytest
requests
pandas 